Added facility to specify nodes, argument, and documentation
Needed to use this script to break down use and availability of
gpus over different nodes (so we can spot if any nodes are being
hogged / are broken easily e.g. if a number of gpus on a given node
are never used, it's likely broken.

I also changed the output format such that we can supress the header
(this is to allow it to be called in a for loop to make output per
node for instance.

I have not investigated the sinfo commands. Is there a reason that
the number of used plus available does not equal total?